Minor fixes

+++ src/main/kotlin/no/nav/pensjon/simulator/tjenestepensjon/pre2025/simulering/client/spk/acl/PrognoseSpecMapper.kt
@@ -87,7 +87,7 @@ object PrognoseSpecMapper {
         OpptjeningsperiodeDto(
             datoFom = spec.fom,
             datoTom = spec.tom,
-            stillingsprosent = spec.stillingsprosent.toDouble(),
+            stillingsprosent = spec.stillingsprosent,
             aldersgrense = spec.aldersgrense,
             faktiskHovedlonn = spec.faktiskHovedloenn,
             stillingsuavhengigTilleggslonn = spec.stillingsuavhengigTilleggsloenn
